Polish code.
