+++ Cargo.toml
@@ -17,6 +17,5 @@ members = [
     "components/merkledb",
     "components/proto",
     "components/proto-derive",
-    "components/derive-test",
 ]
 exclude = [ "exonum/fuzz" ]
