--- components/derive-test/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "derive-test"
-version = "0.12.0"
-edition = "2018"
-authors = ["The Exonum Team <exonum@bitfury.com>"]
-homepage = "https://exonum.com/"
-repository = "https://github.com/exonum/exonum"
-license = "Apache-2.0"
-keywords = ["exonum", "cryptography", "library"]
-categories = ["cryptography", "hashing"]
-description = "Exonum validator keys management."
-
-[dependencies]
-rand = "0.7"
-failure = "0.1.5"
-pwbox = "0.2.1"
-secret-tree = "0.2.0"
-serde = "1.0.10"
-serde_derive = "1.0.64"
-toml = "0.5.0"
-protobuf = { version = "2.8.1", features = ["with-serde"] }
-
-exonum-crypto = { version = "0.12.0", path = "../crypto" }
-exonum-proto = { version = "0.12.0", path = "../proto" }
-exonum-proto-derive = { version = "0.12.0", path = "../proto-derive" }
-
-[dev-dependencies]
-tempdir = "0.3.7"
-hex = "0.4"
-
-[build-dependencies]
-exonum-build = { version = "0.12.0", path = "../build" }
