+++ components/proto-derive/src/lib.rs
@@ -21,30 +21,23 @@
 extern crate proc_macro;
 
 mod pb_convert;
-mod pb_convert2;
 
 use darling::FromMeta;
 use proc_macro::TokenStream;
 use quote::ToTokens;
 use syn::{Attribute, NestedMeta};
 
-/// Derive `ProtobufConvert` trait.
+/// ProtobufConvert attribute macros.
 ///
 /// Attributes:
 ///
 /// ## Required
 ///
-/// * `#[exonum(pb = "path")]`
+/// * `#[protobuf_convert(source = "path")]`
 ///
 /// Path is the name of the corresponding protobuf generated struct.
 ///
-/// ## Optional
-///
-/// * `#[exonum(crate = "path")]`
-///
-/// Prefix of the `exonum` crate(usually "crate" or "exonum").
-///
-/// * `#[exonum(serde_pb_convert)]`
+/// * `#[protobuf_convert(source = "path", serde_pb_convert)]`
 ///
 /// Implement `serde::{Serialize, Deserialize}` using structs that were generated with
 /// protobuf.
@@ -65,14 +58,9 @@ use syn::{Attribute, NestedMeta};
 ///     // ...
 /// }
 /// ```
-#[proc_macro_derive(ProtobufConvert, attributes(exonum))]
-pub fn generate_protobuf_convert(input: TokenStream) -> TokenStream {
-    pb_convert::implement_protobuf_convert(input)
-}
-
 #[proc_macro_attribute]
 pub fn protobuf_convert(attrs: TokenStream, input: TokenStream) -> TokenStream {
-    pb_convert2::implement_protobuf_convert(attrs, input)
+    pb_convert::implement_protobuf_convert(attrs, input)
 }
 
 pub(crate) fn find_exonum_meta(args: &[Attribute]) -> Option<NestedMeta> {
